+++ src/components/FindPinPointLocationStack/FindPinPointLocation.tsx
@@ -1,14 +1,13 @@
 import { TuseState } from '@types';
-import React, { useState, useEffect } from 'react'
-import { Alert} from 'react-native'
+import React from 'react'
 import { GooglePlaceSearchBar } from '../../atoms/GoogleMap';
 
 import { PROVIDER_GOOGLE, Marker } from 'react-native-maps';
 import MapView from 'react-native-maps';
 import { GooglePlaceDetail, GooglePlaceData } from 'react-native-google-places-autocomplete';
-import * as Location from 'expo-location';
 import { Container, Box, Googleplace } from '../../atoms/elements/layouts';
 import { SubTitle } from '../../atoms/elements/texts';
+import { Text } from 'react-native-elements';
 
 const GOOGLE_PLACES_API_KEY = 'AIzaSyA-4i3FV1KLsJbsyVySpYi4YIwxIkEXFlw';
 
@@ -21,41 +20,6 @@ interface Props {
 const FindPinPointLocation = (props: Props) => {
     const [latitude, setLatitude] = props.useLatitude;
     const [longitude, setLongitude] = props.useLongitude;
-    const [userLocation, setUserLocation] = useState(
-        {
-            loaded: true,
-            coordinates: {
-                lat: 0,
-                lng: 0,
-                latDelta: 0.005,
-                lngDelta: 0.005
-            }
-        }
-    )
-
-
-    const getLocation = async () => {
-        try {
-            await Location.requestPermissionsAsync();
-            const { coords } = await Location.getCurrentPositionAsync();
-            setUserLocation({
-                loaded: true,
-                coordinates: {
-                    lat: coords.latitude,
-                    lng: coords.longitude,
-                    latDelta: 0.005,
-                    lngDelta: 0.005
-                }
-            });
-
-
-        } catch (error) {
-            Alert.alert("Can't find you");
-        }
-
-
-    }
-
 
     const onPressMap = (e: { nativeEvent: { coordinate: any; } }) => {
         const { coordinate } = e.nativeEvent
@@ -64,30 +28,22 @@ const FindPinPointLocation = (props: Props) => {
     }
 
 
-    const getPlaceDetails = (data: GooglePlaceData, detail: GooglePlaceDetail|null) => {
+    const getPlaceDetails = (data: GooglePlaceData, detail: GooglePlaceDetail | null) => {
         if (detail === null) return;
-        //console.log(detail?.geometry.location)
+
         const { geometry: { location } } = detail
         setLatitude(location.lat)
         setLongitude(location.lng)
     }
 
-    useEffect(() => {
-        getLocation();
-
-
-    }, []);
-
-
 
 
     return (
-
         <Container>
-                <Box>
+            <Box>
                 <SubTitle>위치 검색</SubTitle>
-                </Box>
-                
+            </Box>
+
             <Googleplace>
                 <GooglePlaceSearchBar
                     placeholder='장소 검색'
@@ -99,40 +55,24 @@ const FindPinPointLocation = (props: Props) => {
                     onFail={(error) => console.log(error)}
                     fetchDetails={true}
                 />
-
             </Googleplace>
+
             <Box>
                 <SubTitle>지도에서 보기</SubTitle>
-                </Box>
+            </Box>
 
-            {(latitude && longitude) ?
-                
-                <MapView
-                    style={{ flex: 1 }}
-                    provider={PROVIDER_GOOGLE}
-                    region={{ latitude: latitude, longitude: longitude, latitudeDelta: 0.004, longitudeDelta: 0.004 }}
-                    showsUserLocation={true}
-                    showsMyLocationButton={true}
-                    onPress={onPressMap}
-                >
-                    <Marker coordinate={{ latitude: latitude, longitude: longitude }} />
-                </MapView>
-                : <MapView style={{ flex: 1 }}
-                    provider={PROVIDER_GOOGLE}
-                    region={{ latitude: userLocation.coordinates.lat, longitude: userLocation.coordinates.lng, latitudeDelta: 0.004, longitudeDelta: 0.004 }}
-                    showsUserLocation={true}
-                    showsMyLocationButton={true}
-                    onPress={onPressMap}>
-                </MapView>
-
-
-
-
-            }
+            {!longitude && !latitude && <Text>유저 위치 찾는 중..</Text>}
+            <MapView
+                style={{ flex: 1 }}
+                provider={PROVIDER_GOOGLE}
+                region={{ latitude: latitude, longitude: longitude, latitudeDelta: 0.004, longitudeDelta: 0.004 }}
+                showsUserLocation={true}
+                showsMyLocationButton={true}
+                onPress={onPressMap}
+            >
+                <Marker coordinate={{ latitude: latitude, longitude: longitude }} />
+            </MapView>
         </Container>
-
-
-
     )
 }
